[DEV] continue rework

+++ tools/testbed/scenes/cubes/CubesScene.cpp
@@ -1,26 +1,26 @@
 /********************************************************************************
-* ReactPhysics3D physics library, http://www.reactphysics3d.com                 *
-* Copyright (c) 2010-2016 Daniel Chappuis                                       *
+* ReactPhysics3D physics library, http://www.reactphysics3d.com				 *
+* Copyright (c) 2010-2016 Daniel Chappuis									   *
 *********************************************************************************
-*                                                                               *
+*																			   *
 * This software is provided 'as-is', without any express or implied warranty.   *
 * In no event will the authors be held liable for any damages arising from the  *
-* use of this software.                                                         *
-*                                                                               *
-* Permission is granted to anyone to use this software for any purpose,         *
-* including commercial applications, and to alter it and redistribute it        *
-* freely, subject to the following restrictions:                                *
-*                                                                               *
+* use of this software.														 *
+*																			   *
+* Permission is granted to anyone to use this software for any purpose,		 *
+* including commercial applications, and to alter it and redistribute it		*
+* freely, subject to the following restrictions:								*
+*																			   *
 * 1. The origin of this software must not be misrepresented; you must not claim *
-*    that you wrote the original software. If you use this software in a        *
-*    product, an acknowledgment in the product documentation would be           *
-*    appreciated but is not required.                                           *
-*                                                                               *
-* 2. Altered source versions must be plainly marked as such, and must not be    *
-*    misrepresented as being the original software.                             *
-*                                                                               *
-* 3. This notice may not be removed or altered from any source distribution.    *
-*                                                                               *
+*	that you wrote the original software. If you use this software in a		*
+*	product, an acknowledgment in the product documentation would be		   *
+*	appreciated but is not required.										   *
+*																			   *
+* 2. Altered source versions must be plainly marked as such, and must not be	*
+*	misrepresented as being the original software.							 *
+*																			   *
+* 3. This notice may not be removed or altered from any source distribution.	*
+*																			   *
 ********************************************************************************/
 
 // Libraries
@@ -32,167 +32,167 @@ using namespace cubesscene;
 
 // Constructor
 CubesScene::CubesScene(const std::string& name)
-      : SceneDemo(name, SCENE_RADIUS) {
+	  : SceneDemo(name, SCENE_RADIUS) {
 
-    // Compute the radius and the center of the scene
-    openglframework::Vector3 center(0, 5, 0);
+	// Compute the radius and the center of the scene
+	openglframework::Vector3 center(0, 5, 0);
 
-    // Set the center of the scene
-    setScenePosition(center, SCENE_RADIUS);
+	// Set the center of the scene
+	setScenePosition(center, SCENE_RADIUS);
 
-    // Gravity vector in the dynamics world
-    rp3d::Vector3 gravity(0, rp3d::decimal(-9.81), 0);
+	// Gravity vector in the dynamics world
+	rp3d::Vector3 gravity(0, rp3d::float(-9.81), 0);
 
-    // Create the dynamics world for the physics simulation
-    m_dynamicsWorld = new rp3d::DynamicsWorld(gravity);
+	// Create the dynamics world for the physics simulation
+	m_dynamicsWorld = new rp3d::DynamicsWorld(gravity);
 
-    // Set the number of iterations of the constraint solver
-    m_dynamicsWorld->setNbIterationsVelocitySolver(15);
+	// Set the number of iterations of the constraint solver
+	m_dynamicsWorld->setNbIterationsVelocitySolver(15);
 
-    float radius = 2.0f;
+	float radius = 2.0f;
 
-    // Create all the cubes of the scene
-    for (int i=0; i<NB_CUBES; i++) {
+	// Create all the cubes of the scene
+	for (int32_t i=0; i<NB_CUBES; i++) {
 
-        // Position of the cubes
-        float angle = i * 30.0f;
-        openglframework::Vector3 position(radius * cos(angle),
-                                          30 + i * (BOX_SIZE.y + 0.3f),
-                                          0);
+		// Position of the cubes
+		float angle = i * 30.0f;
+		openglframework::Vector3 position(radius * cos(angle),
+										  30 + i * (BOX_SIZE.y + 0.3f),
+										  0);
 
-        // Create a cube and a corresponding rigid in the dynamics world
-        Box* cube = new Box(BOX_SIZE, position , BOX_MASS, m_dynamicsWorld);
+		// Create a cube and a corresponding rigid in the dynamics world
+		Box* cube = new Box(BOX_SIZE, position , BOX_MASS, m_dynamicsWorld);
 
-        // Set the box color
-        cube->setColor(mDemoColors[i % mNbDemoColors]);
-        cube->setSleepingColor(mRedColorDemo);
+		// Set the box color
+		cube->setColor(mDemoColors[i % mNbDemoColors]);
+		cube->setSleepingColor(mRedColorDemo);
 
-        // Change the material properties of the rigid body
-        rp3d::Material& material = cube->getRigidBody()->getMaterial();
-        material.setBounciness(rp3d::decimal(0.4));
+		// Change the material properties of the rigid body
+		rp3d::Material& material = cube->getRigidBody()->getMaterial();
+		material.setBounciness(rp3d::float(0.4));
 
-        // Add the box the list of box in the scene
-        mBoxes.push_back(cube);
-    }
+		// Add the box the list of box in the scene
+		mBoxes.push_back(cube);
+	}
 
-    // Create the floor
-    openglframework::Vector3 floorPosition(0, 0, 0);
-    mFloor = new Box(FLOOR_SIZE, floorPosition, FLOOR_MASS, m_dynamicsWorld);
-    mFloor->setColor(mGreyColorDemo);
-    mFloor->setSleepingColor(mGreyColorDemo);
+	// Create the floor
+	openglframework::Vector3 floorPosition(0, 0, 0);
+	mFloor = new Box(FLOOR_SIZE, floorPosition, FLOOR_MASS, m_dynamicsWorld);
+	mFloor->setColor(mGreyColorDemo);
+	mFloor->setSleepingColor(mGreyColorDemo);
 
-    // The floor must be a static rigid body
-    mFloor->getRigidBody()->setType(rp3d::STATIC);
+	// The floor must be a static rigid body
+	mFloor->getRigidBody()->setType(rp3d::STATIC);
 
-    // Change the material properties of the floor rigid body
-    rp3d::Material& material = mFloor->getRigidBody()->getMaterial();
-    material.setBounciness(rp3d::decimal(0.3));
+	// Change the material properties of the floor rigid body
+	rp3d::Material& material = mFloor->getRigidBody()->getMaterial();
+	material.setBounciness(rp3d::float(0.3));
 
-    // Get the physics engine parameters
-    mEngineSettings.isGravityEnabled = m_dynamicsWorld->isGravityEnabled();
-    rp3d::Vector3 gravityVector = m_dynamicsWorld->getGravity();
-    mEngineSettings.gravity = openglframework::Vector3(gravityVector.x, gravityVector.y, gravityVector.z);
-    mEngineSettings.isSleepingEnabled = m_dynamicsWorld->isSleepingEnabled();
-    mEngineSettings.sleepLinearVelocity = m_dynamicsWorld->getSleepLinearVelocity();
-    mEngineSettings.sleepAngularVelocity = m_dynamicsWorld->getSleepAngularVelocity();
-    mEngineSettings.nbPositionSolverIterations = m_dynamicsWorld->getNbIterationsPositionSolver();
-    mEngineSettings.nbVelocitySolverIterations = m_dynamicsWorld->getNbIterationsVelocitySolver();
-    mEngineSettings.timeBeforeSleep = m_dynamicsWorld->getTimeBeforeSleep();
+	// Get the physics engine parameters
+	mEngineSettings.isGravityEnabled = m_dynamicsWorld->isGravityEnabled();
+	rp3d::Vector3 gravityVector = m_dynamicsWorld->getGravity();
+	mEngineSettings.gravity = openglframework::Vector3(gravityVector.x, gravityVector.y, gravityVector.z);
+	mEngineSettings.isSleepingEnabled = m_dynamicsWorld->isSleepingEnabled();
+	mEngineSettings.sleepLinearVelocity = m_dynamicsWorld->getSleepLinearVelocity();
+	mEngineSettings.sleepAngularVelocity = m_dynamicsWorld->getSleepAngularVelocity();
+	mEngineSettings.nbPositionSolverIterations = m_dynamicsWorld->getNbIterationsPositionSolver();
+	mEngineSettings.nbVelocitySolverIterations = m_dynamicsWorld->getNbIterationsVelocitySolver();
+	mEngineSettings.timeBeforeSleep = m_dynamicsWorld->getTimeBeforeSleep();
 }
 
 // Destructor
 CubesScene::~CubesScene() {
 
-    // Destroy all the cubes of the scene
-    for (std::vector<Box*>::iterator it = mBoxes.begin(); it != mBoxes.end(); ++it) {
+	// Destroy all the cubes of the scene
+	for (std::vector<Box*>::iterator it = mBoxes.begin(); it != mBoxes.end(); ++it) {
 
-        // Destroy the corresponding rigid body from the dynamics world
-        m_dynamicsWorld->destroyRigidBody((*it)->getRigidBody());
+		// Destroy the corresponding rigid body from the dynamics world
+		m_dynamicsWorld->destroyRigidBody((*it)->getRigidBody());
 
-        // Destroy the cube
-        delete (*it);
-    }
+		// Destroy the cube
+		delete (*it);
+	}
 
-    // Destroy the rigid body of the floor
-    m_dynamicsWorld->destroyRigidBody(mFloor->getRigidBody());
+	// Destroy the rigid body of the floor
+	m_dynamicsWorld->destroyRigidBody(mFloor->getRigidBody());
 
-    // Destroy the floor
-    delete mFloor;
+	// Destroy the floor
+	delete mFloor;
 
-    // Destroy the dynamics world
-    delete m_dynamicsWorld;
+	// Destroy the dynamics world
+	delete m_dynamicsWorld;
 }
 
 // Update the physics world (take a simulation step)
 void CubesScene::updatePhysics() {
 
-    // Update the physics engine parameters
-    m_dynamicsWorld->setIsGratityEnabled(mEngineSettings.isGravityEnabled);
-    rp3d::Vector3 gravity(mEngineSettings.gravity.x, mEngineSettings.gravity.y,
-                                     mEngineSettings.gravity.z);
-    m_dynamicsWorld->setGravity(gravity);
-    m_dynamicsWorld->enableSleeping(mEngineSettings.isSleepingEnabled);
-    m_dynamicsWorld->setSleepLinearVelocity(mEngineSettings.sleepLinearVelocity);
-    m_dynamicsWorld->setSleepAngularVelocity(mEngineSettings.sleepAngularVelocity);
-    m_dynamicsWorld->setNbIterationsPositionSolver(mEngineSettings.nbPositionSolverIterations);
-    m_dynamicsWorld->setNbIterationsVelocitySolver(mEngineSettings.nbVelocitySolverIterations);
-    m_dynamicsWorld->setTimeBeforeSleep(mEngineSettings.timeBeforeSleep);
+	// Update the physics engine parameters
+	m_dynamicsWorld->setIsGratityEnabled(mEngineSettings.isGravityEnabled);
+	rp3d::Vector3 gravity(mEngineSettings.gravity.x, mEngineSettings.gravity.y,
+									 mEngineSettings.gravity.z);
+	m_dynamicsWorld->setGravity(gravity);
+	m_dynamicsWorld->enableSleeping(mEngineSettings.isSleepingEnabled);
+	m_dynamicsWorld->setSleepLinearVelocity(mEngineSettings.sleepLinearVelocity);
+	m_dynamicsWorld->setSleepAngularVelocity(mEngineSettings.sleepAngularVelocity);
+	m_dynamicsWorld->setNbIterationsPositionSolver(mEngineSettings.nbPositionSolverIterations);
+	m_dynamicsWorld->setNbIterationsVelocitySolver(mEngineSettings.nbVelocitySolverIterations);
+	m_dynamicsWorld->setTimeBeforeSleep(mEngineSettings.timeBeforeSleep);
 
-    // Take a simulation step
-    m_dynamicsWorld->update(mEngineSettings.timeStep);
+	// Take a simulation step
+	m_dynamicsWorld->update(mEngineSettings.timeStep);
 }
 
 // Update the scene
 void CubesScene::update() {
 
-    SceneDemo::update();
+	SceneDemo::update();
 
-    // Update the position and orientation of the boxes
-    for (std::vector<Box*>::iterator it = mBoxes.begin(); it != mBoxes.end(); ++it) {
+	// Update the position and orientation of the boxes
+	for (std::vector<Box*>::iterator it = mBoxes.begin(); it != mBoxes.end(); ++it) {
 
-        // Update the transform used for the rendering
-        (*it)->updateTransform(mInterpolationFactor);
-    }
+		// Update the transform used for the rendering
+		(*it)->updateTransform(mInterpolationFactor);
+	}
 
-    mFloor->updateTransform(mInterpolationFactor);
+	mFloor->updateTransform(mInterpolationFactor);
 }
 
 // Render the scene in a single pass
 void CubesScene::renderSinglePass(Shader& shader, const openglframework::Matrix4& worldToCameraMatrix) {
 
-    // Bind the shader
-    shader.bind();
+	// Bind the shader
+	shader.bind();
 
-    // Render all the cubes of the scene
-    for (std::vector<Box*>::iterator it = mBoxes.begin(); it != mBoxes.end(); ++it) {
-        (*it)->render(shader, worldToCameraMatrix);
-    }
+	// Render all the cubes of the scene
+	for (std::vector<Box*>::iterator it = mBoxes.begin(); it != mBoxes.end(); ++it) {
+		(*it)->render(shader, worldToCameraMatrix);
+	}
 
-    // Render the floor
-    mFloor->render(shader, worldToCameraMatrix);
+	// Render the floor
+	mFloor->render(shader, worldToCameraMatrix);
 
-    // Unbind the shader
-    shader.unbind();
+	// Unbind the shader
+	shader.unbind();
 }
 
 // Reset the scene
 void CubesScene::reset() {
 
-    float radius = 2.0f;
+	float radius = 2.0f;
 
-    for (int i=0; i<NB_CUBES; i++) {
+	for (int32_t i=0; i<NB_CUBES; i++) {
 
-        // Position of the cubes
-        float angle = i * 30.0f;
-        openglframework::Vector3 position(radius * cos(angle),
-                                          10 + i * (BOX_SIZE.y + 0.3f),
-                                          0);
+		// Position of the cubes
+		float angle = i * 30.0f;
+		openglframework::Vector3 position(radius * cos(angle),
+										  10 + i * (BOX_SIZE.y + 0.3f),
+										  0);
 
-        // Initial position and orientation of the rigid body
-        rp3d::Vector3 initPosition(position.x, position.y, position.z);
-        rp3d::Quaternion initOrientation = rp3d::Quaternion::identity();
-        rp3d::Transform transform(initPosition, initOrientation);
+		// Initial position and orientation of the rigid body
+		rp3d::Vector3 initPosition(position.x, position.y, position.z);
+		rp3d::Quaternion initOrientation = rp3d::Quaternion::identity();
+		rp3d::Transform transform(initPosition, initOrientation);
 
-        mBoxes[i]->resetTransform(transform);
-    }
+		mBoxes[i]->resetTransform(transform);
+	}
 }
